requirements.txt
Django==1.6.1
South==0.8.4
dj-database-url==0.2.2
dj-static==0.0.5
django-toolbelt==0.0.1
django-vanilla-views==1.0.2
facebook-sdk==0.4.0
gunicorn==18.0
psycopg2==2.5.1
static==0.4
wsgiref==0.1.2
